# server/docs/openapi.yaml
openapi: 3.0.0
info:
  description: REST API for the NDHU campus second-hand textbook marketplace.
  title: NDHU Second-Hand Book Store API
  termsOfService: http://swagger.io/terms/
  contact:
    name: SHBS Dev Team
  license:
    name: MIT
  version: "1.0"
paths:
...
  "/listings/{listingId}/messages":
    get:
      security:
        - BearerAuth: []
      description: Returns the conversation between the authenticated user and the
        other party for a given listing. Marks received messages as read.
      tags:
        - Messages
      summary: List messages for a listing
      operationId: getMessages
      parameters:
        - description: Listing ID
          name: listingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/model.MessageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
    post:
      security:
        - BearerAuth: []
      description: Sends a message from the authenticated user to the listing seller.
        Creates a notification for the receiver.
      tags:
        - Messages
      summary: Send a message about a listing
      operationId: sendMessage
      parameters:
        - description: Listing ID
          name: listingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/model.SwaggerSendMessageRequest"
        description: Message body
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.MessageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
  /messages/conversations:
    get:
      security:
        - BearerAuth: []
      description: Returns one summary entry per unique (listing, other_user)
        conversation pair the authenticated user is part of.
      tags:
        - Messages
      summary: List conversation threads
      operationId: getConversations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/model.ConversationResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
  /messages/unread-count:
    get:
      security:
        - BearerAuth: []
      description: Returns the number of unread messages for the authenticated user.
      tags:
        - Messages
      summary: Unread message count
      operationId: getUnreadMessageCount
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
  "/messages/{id}/read":
    patch:
      security:
        - BearerAuth: []
      description: Marks a single message as read. Only the receiver may mark a
        message as read.
      tags:
        - Messages
      summary: Mark message as read
      operationId: markMessageAsRead
      parameters:
        - description: Message ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
...
servers:
  - url: /api/v1
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    model.ConversationResponse:
      type: object
      properties:
        last_message:
          type: string
        last_message_at:
          type: string
        listing_id:
          type: string
        listing_title:
          type: string
        other_user_id:
          type: string
        other_user_name:
          type: string
        unread_count:
          type: integer
...
    model.MessageResponse:
      type: object
      properties:
        body:
          type: string
        created_at:
          type: string
        id:
          type: string
        is_read:
          type: boolean
        listing_id:
          type: string
        receiver_id:
          type: string
        sender_id:
          type: string
        sender_name:
          type: string
...
    model.SwaggerErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    model.SwaggerSendMessageRequest:
      type: object
      properties:
        body:
          type: string
...